+++ docs/pages/release_notes.md
@@ -46,6 +46,8 @@ This is a {{ site.pmd.release_type }} release.
     ```
 
 ### 🐛️ Fixed Issues
+* java-bestpractices
+    * [#6984](https://github.com/pmd/pmd/issues/6984): \[java] ExhaustiveSwitchHasDefault: Improve violation message to list missing cases
 * java-codestyle
     * [#5732](https://github.com/pmd/pmd/issues/5732): \[java] UnnecessaryCast false positive with package private methods
 * java-errorprone